Move new classes to own files. Add comments
